+++ src/main/java/com/bookwise/bookwise/service/impl/UserServiceImpl.java
@@ -1,33 +1,25 @@
 package com.bookwise.bookwise.service.impl;
 
-import com.bookwise.bookwise.constants.JWTConstants;
 import com.bookwise.bookwise.dto.user.RegisterRequestDTO;
 import com.bookwise.bookwise.dto.user.UserDTO;
 import com.bookwise.bookwise.entity.User;
+import com.bookwise.bookwise.exception.ResourceAlreadyExistsException;
 import com.bookwise.bookwise.mapper.UserMapper;
+import com.bookwise.bookwise.repository.IssuanceRepository;
 import com.bookwise.bookwise.repository.UserRepository;
+import com.bookwise.bookwise.service.ISMSService;
 import com.bookwise.bookwise.service.IUserService;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
+import com.bookwise.bookwise.utils.PasswordGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.env.Environment;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
@@ -35,7 +27,9 @@ import java.util.stream.Collectors;
 public class UserServiceImpl implements IUserService {
 
     private final UserRepository userRepository;
+    private final ISMSService ismsService;
     private final PasswordEncoder passwordEncoder;
+    private final IssuanceRepository issuanceRepository;
     private final Environment env;
 
     @Override
@@ -58,18 +52,6 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public UserDTO getUserByEmail(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(
-                () -> new UsernameNotFoundException("User not found for " + email)
-        );
-
-        UserDTO userDTO = UserMapper.mapToUserDTO(user, new UserDTO());
-
-        return userDTO;
-
-    }
-
-    @Override
     public UserDTO getUserByMobile(String mobileNumber) {
         User user = userRepository.findByMobileNumber(mobileNumber).orElseThrow(
                 () -> new UsernameNotFoundException("User not found for " + mobileNumber)
@@ -82,31 +64,15 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public List<UserDTO> getAllUsers() {
-        List<User> userList = userRepository.findAll();
-        List<UserDTO> userDTOList = new ArrayList<>();
-
-        userList.forEach(user -> userDTOList.add(UserMapper.mapToUserDTO(user, new UserDTO())));
-        return userDTOList;
-    }
-
-    @Override
-    public Long getUserCount() {
-        Long userCount = userRepository.count();
-        return userCount;
-    }
-
-    @Override
-    public UserDTO deleteUserByEmail(String email) {
+    public UserDTO getUserByEmail(String email) {
         User user = userRepository.findByEmail(email).orElseThrow(
                 () -> new UsernameNotFoundException("User not found for " + email)
         );
 
-        userRepository.deleteById(user.getId());
-
         UserDTO userDTO = UserMapper.mapToUserDTO(user, new UserDTO());
 
         return userDTO;
+
     }
 
     @Override
@@ -115,6 +81,13 @@ public class UserServiceImpl implements IUserService {
                 () -> new UsernameNotFoundException("User not found for " + mobileNumber)
         );
 
+        boolean isBookIssued = issuanceRepository.existsByUserIdAndStatus(user.getId(), "Issued");
+
+        if (isBookIssued) {
+            throw new IllegalStateException("The user has issued some books and cannot be deleted.");
+        }
+
+        issuanceRepository.deleteAllByUserIn(Collections.singletonList(user));
         userRepository.deleteById(user.getId());
 
         UserDTO userDTO = UserMapper.mapToUserDTO(user, new UserDTO());
@@ -124,12 +97,41 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public UserDTO registerUser(RegisterRequestDTO registerRequestDTO) {
+
+        Optional<User> optionalUser = userRepository.findByMobileNumber(registerRequestDTO.getMobileNumber());
+        if (optionalUser.isPresent()) {
+            throw new ResourceAlreadyExistsException("User already exists for mobile no. " + registerRequestDTO.getMobileNumber());
+        }
+
+        optionalUser = userRepository.findByEmail(registerRequestDTO.getEmail());
+        if (optionalUser.isPresent()) {
+            throw new ResourceAlreadyExistsException("User already exists for email " + registerRequestDTO.getEmail());
+        }
+
         User user = UserMapper.mapToUser(registerRequestDTO, new User());
-        user.setPassword(passwordEncoder.encode(registerRequestDTO.getPassword()));
+
+        String randomPassword = PasswordGenerator.generatePassword(10);
+        System.out.println("PASSWORD -> " + randomPassword);
+
+        user.setPassword(passwordEncoder.encode(randomPassword));
         if (user.getRole() == null || user.getRole().isEmpty()) {
             user.setRole("ROLE_USER");
         }
+
         User savedUser = userRepository.save(user);
+
+        String message = String.format( "\nWelcome %s\n" +
+                                        "You have successfully registered to BookWise\n" +
+                                        "These are your login credentials\n" +
+                                        "Username: %s (OR) %s\n" +
+                                        "Password: %s",
+                savedUser.getName(),
+                savedUser.getMobileNumber(),
+                savedUser.getEmail(),
+                randomPassword);
+
+        ismsService.sendSms(savedUser.getMobileNumber(), message);
+
         UserDTO userDTO = UserMapper.mapToUserDTO(savedUser, new UserDTO());
         return  userDTO;
     }
@@ -140,8 +142,32 @@ public class UserServiceImpl implements IUserService {
                 () -> new UsernameNotFoundException("User not found for mobile no. " + mobileNumber)
         );
 
+        Optional<User> optionalUser = userRepository.findByMobileNumber(registerRequestDTO.getMobileNumber());
+        if (optionalUser.isPresent()) {
+            User otherUser = optionalUser.get();
+            if (otherUser.getId() != user.getId()) {
+                throw new ResourceAlreadyExistsException("User already exists for mobile no. " + registerRequestDTO.getMobileNumber());
+            }
+        }
+
+        optionalUser = userRepository.findByEmail(registerRequestDTO.getEmail());
+        if (optionalUser.isPresent()) {
+            User otherUser = optionalUser.get();
+            if (otherUser.getId() != user.getId()) {
+                throw new ResourceAlreadyExistsException("User already exists for email " + registerRequestDTO.getEmail());
+            }
+        }
+
         user = UserMapper.mapToUser(registerRequestDTO, user);
-        user.setPassword(passwordEncoder.encode(registerRequestDTO.getPassword()));
+
+        if (registerRequestDTO.getPassword() != null && registerRequestDTO.getPassword().length() >= 3) {
+            String encodedPassword = registerRequestDTO.getPassword();
+            byte[] decodedBytes = Base64.getDecoder().decode(encodedPassword);
+            String decodedPassword = new String(decodedBytes);
+            registerRequestDTO.setPassword(decodedPassword);
+            user.setPassword(passwordEncoder.encode(registerRequestDTO.getPassword()));
+        }
+
         if (user.getRole() == null || user.getRole().isEmpty()) {
             user.setRole("ROLE_USER");
         }
@@ -154,33 +180,4 @@ public class UserServiceImpl implements IUserService {
 
     }
 
-    @Override
-    public UserDTO getUserByToken(String jwt) {
-        String secret = env.getProperty(JWTConstants.JWT_SECRET_KEY, JWTConstants.JWT_SECRET_DEFAULT_VALUE);
-        SecretKey secretKey = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
-        Claims claims = Jwts.parser().verifyWith(secretKey)
-                .build().parseSignedClaims(jwt).getPayload();
-        String username = String.valueOf(claims.get("username"));
-
-        if (username.contains("@")) {
-//                            email
-            User user = userRepository.findByEmail(username).orElseThrow(
-                    () -> new BadCredentialsException("Invalid Token received!")
-            );
-
-            UserDTO userDTO = UserMapper.mapToUserDTO(user, new UserDTO());
-            userDTO.setToken(jwt);
-            return  userDTO;
-        } else {
-//                            mobile
-            User user = userRepository.findByMobileNumber(username).orElseThrow(
-                    () -> new BadCredentialsException("Invalid Token received!")
-            );
-
-            UserDTO userDTO = UserMapper.mapToUserDTO(user, new UserDTO());
-            userDTO.setToken(jwt);
-            return  userDTO;
-        }
-
-    }
 }
